changed expected filenames; renamed DefaultTemplateName (again)
Meta files are now treated as Asset files if not named "defaults" or
"meta". This is to enforce a better standard file naming scheme
throughout projects and make them less confusing to parse and to make
it easier to document how to build a project.

I decided "default" was a better default DefaultTemplateName than
"_root".