added Assets type to Page (for .Assets).
This allows users to navigate assets better. It has '.All' for all
asset paths (a []string as before). '.All' has been split into
'.Image', '.Video', '.Audio' and '.Misc', each being a []*string
(each entry points to an '.All' entry) of the respective types.

Implemented this because I found myself wanting to display a gallery
of images and link a .zip of all the displayed images. I could have
added the .zip to the AssetDir but meant moving it out of that
gallery project folder.